Update RP2040 install instructions
Adds a link to the needed RP2040 Arduino core, since the one downloaded from Arduino IDE's "Boards Manager" does not work as-is. Removes the notice to make sure the version is at least 3.6.1, since that code is currently on 4.4.4 at the time of writing, so any version downloaded should be adequate for use.